auth/requirements.txt
redis==4.6.0
fastapi==0.100.0
pydantic==2.1.1
pydantic-settings==2.0.2
pyjwt==2.8.0
uvicorn==0.22.0
SQLAlchemy==2.0.19
psycopg==3.1.9
python-dotenv==1.0.0
alembic==1.11.2
psycopg2-binary==2.9.7
pytest==7.4.0
aiohttp==3.8.5
gunicorn==21.2.0
asyncpg==0.28.0
grpcio-tools==1.57.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.19.0
fastapi-limiter==0.1.5